Fix typo in revert-state-upon-exception equation 91
This equation previously read:

    sigma' = sigma1 if o==emptyset, else sigma**

I think that "o" is never the empty set: my reading of eqs 92, 103, 109, 110, 113, and 118 leads me to believe that o will be:

* () if a Z=True "exceptional halt" occurs (out-of-gas, bad-instruction,
  stack-underflow, bad-jumpdest, call-stack-overflow)
* () if STOP or SUICIDE instructions are executed
* () if RETURN is executed with length-to-return = 0
* (stuff) is RETURN is executed with length-to-return > 0

And I think that the revert-to-sigma1 behavior is intended to only occur when an exceptional halt occurs. If we test `sigma**` instead of "o", I think we'll get the intended behavior.