Remove the modulo 2^{256} effect in the memory size computation
This commit adds notes to MLOAD, MSTORE and MSTORE8 specification
saying the modulo is not considered when computing the new memory
size.

Before this change, since the beginning of the table contains
`All arithmetic is modulo $2^{256}$ unless otherwise noted.`,
the memory size computation was affected by the modulo computation too.
For instance, MLOAD with the biggest uint256 didn't change the memory size
most of the time.