Correct transaction execution gas used equation
This commit seeks to address a disconnect between the Yellow Paper
specification and what is implemented in the Ethereum clients in
practice for calculating the correct transaction receipt field.

Currently the Yellow Paper specifies that the total gas by a
transaction is the difference between the transaction gas limit
and the gas remaining from the root-level message call or contract
creation within the transaction.

This neglects the fact that the transaction sender can be refunded
additional gas after the root-level message call or contract creation
has completed, which is problematic because this is used to compute
the cumulative gas used field for the transaction's corresponding
receipt in practice.