Correct self-destruct refund semantics
The current Yellow Paper specification for the self-destruct refunds can
lead to double refunds and is not consistent with the logic clients use
to fulfill refunds in practice.

The crux of the issue is that the current Yellow Paper performs the
self-destruct refund within the SELFDESTRUCT operation of a specific EVM
invocation and each EVM invocation creates its own transaction substate
(rather than having it passed in from its parent). Therefore, although
the transaction substate is checked to see if the account has already
been included in its self-destruct set, that self-destruct set does not
have complete knowledge of the other self-destructions that have
occurred throughout the entire parent transactions execution so far
(e.g. the ancestor and sibling/nibling/etc). Said another way, the EVM
invocation only knows about the self-destructions that have occurred
through the internal message calls and contract creations chains it has
created and an account self-destructed in that chain may have already
been self-destructed outside the scope of it’s invocation.

The cleanest way to handle this issue, which was done by cpp-ethereum
and pyethereum and is reflected in this commit, is to perform the
self-destruct refund accumulation after a transaction’s root message
call or contract creation has completed. At this point, all of the
self-destructs that occur within the transaction have been aggregated,
making it a safe point to calculate this refund. It is worth noting that
geth performs the deduction in the EVM still but incorporates some data
it has added to how it represents world state.