Explicate non-RLP-encodable structures.
Excessive large structures cannot be encoded in RLP, namely byte arrays that
contain 2^64 or more bytes, and lists whose concatenated serialized items
contain 2^64 or more bytes. These restrictions ensure that the first byte of an
encoding is indeed a byte, and that the first bytes of byte arrays vs. list
encodings are disjoint. Also see the encode_length function in the RLP page of
the Ethereum Wiki.

Prior to this commit, the definition of the RLP function in Appendix B did not
explicate these restrictions. Even though these restrictions can be reasonably
inferred from the fact that RLP encodings must be easily decodable, this commit
improves clarity by having the RLP function return an explicit "error" value
when the input structure cannot be encoded. We just need a few more cases in the
equations that define the functions R_b, R_l, and s. The error value is
currently \varnothing, but a different symbol could be used instead.