Clarify the definition of RLP trees.
According to equations (177) and (178), the empty list is () and the empty byte
array is also (). So, if \mathbb{T} is defined as set union, it would contain a
single empty sequence. But the empty byte array and the empty list are encoded
differently, according to equations (180) and (183) -- as the 128 and as 192,
respectively. So \mathbb{T} must contain separate elements for the empty byte
array and for the empty list, which is achieved via a disjoint union. Non-empty
byte arrays and non-empty lists are inherently distinct, since a byte is a
number.