Treat t argument of HP (and f) as boolean.
The paragraph just before eq. (186) says that the second argument, t, of the
function HP, is a boolean value. The function HP is passed the boolean values
true and false as t in eq. (194). However, eq. (187) was treating t as a number,
in the condition 't \neq 0'. Even though booleans are isomorphic to {0, 1}, it's
clearer to change that condition to 't = true'.