Improve appearance of true and false.
In LaTeX's math mode, multi-letter items like true and false should be enclosed
in \mathit{...}, otherwise, each letter is considered a separate item and the
multi-letter item is rendered as an arithmetic product of its letters, with a
bit of additional spacing between the letters.

This commit adds \mathit{...} around all the math-mode occurrences of true and
false.