Pass a bit as the second argument of HP.
Appendix D was passing true and false as the second argument t of HP. However,
HP (and f) in Appendix C treat t as a bit, not as a boolean. Even though bits
and booleans are isomoprhic, it's clearer to change Appendix D to pass bits. The
mention of true and false in the text that precedes eq. (194) has also been
changes to 1 and 0.