Fix mathematical symbols (17 modifications).
These variables (c, o, Id, d) are all arrays. The intention here is to get the length of the array, so apply the norm symbol ( || ) instead of the absolute value symbol ( | ).
The modified content includes the following formulas: 140, 198, 199, 202, 203, 209, 212, 218, 231, 255, 256, 262, 263, 270. And a partial description of the CALL instruction in the appendix "Virtual Machine Specification".